refactor: 폴더 구조 변경
- 한 파일에 있던 UI, 애니메이션, 타입을 각 폴더로 분리
- animations, types, components/sections, common  폴더 생성
