--- src/features/landingPage/FeatureSection.tsx
@@ -1,10 +1,0 @@
-const FeatureSection = () => {
-  return (
-    <div className="min-h-screen w-full flex justify-center items-center">
-      기능 설명 섹션입니다기능 설명 섹션입니다기능 설명 섹션입니다기능 설명 섹션입니다기능 설명
-      섹션입니다
-    </div>
-  );
-};
-
-export default FeatureSection;
