+++ src/features/landingPage/animations/IntroAnimation.ts
@@ -1,0 +1,57 @@
+import gsap from 'gsap';
+import { IntroAnimationProps } from '../types/animation';
+
+export const introAnimation = ({ logoRef, descRef, bgRef, onComplete }: IntroAnimationProps) => {
+  const tl = gsap.timeline();
+
+  if (logoRef.current && descRef.current && bgRef.current) {
+    tl.to(
+      logoRef.current,
+      {
+        y: -40,
+        ease: 'power1.out',
+        zIndex: 30,
+        delay: 1,
+      },
+      'sync'
+    )
+      .fromTo(
+        descRef.current,
+        { opacity: 0, y: 40 },
+        {
+          opacity: 1,
+          y: 0,
+          ease: 'power2.out',
+          zIndex: 30,
+          delay: 1,
+        },
+        'sync'
+      )
+      // 2. 텍스트 색상 하얀색, 배경 색상 검정색으로 변경
+      .add(() => {
+        gsap.to(logoRef.current, { color: 'white' });
+        gsap.to(descRef.current, { color: 'white' });
+        gsap.to(bgRef.current, { backgroundColor: 'black' });
+      }, '+=0.5')
+
+      // 3. 로고 & 멘트 fade out
+      .add(() => {
+        gsap.to(logoRef.current, {
+          y: 0,
+          opacity: 0,
+          scale: 0.7,
+          duration: 1,
+          ease: 'power3.out',
+        });
+        gsap.to(descRef.current, {
+          y: 0,
+          opacity: 0,
+          scale: 0.7,
+          duration: 1,
+          ease: 'power3.out',
+        });
+      }, '+=1')
+
+      .add(onComplete, '+=1');
+  }
+};
