+++ src/features/landingPage/components/common/PurpleCircle.tsx
