+++ src/features/landingPage/components/sections/EarthSection.tsx
