+++ src/features/landingPage/types/animation.ts
@@ -1,0 +1,10 @@
+export type IntroProps = {
+  onComplete: () => void;
+};
+
+export type IntroAnimationProps = {
+  logoRef: React.RefObject<HTMLElement | null>;
+  descRef: React.RefObject<HTMLElement | null>;
+  bgRef: React.RefObject<HTMLElement | null>;
+  onComplete: () => void;
+};
